perf(stuiview): precompute column widths and disable 'EvaluateAllRows' to improve rendering performance
EvaluateAllRows is extremely expensive for large tables. It was enabled
previously as it allows rendering the table such that the column widths
don't jump around as you scroll. Instead, we now compute the widths of
columns on first load of most of the datasources and pad the cells with
proper widths accordingly.